Created a new file isValidFunctions.ts to shared folder to be able to use both in front- and backend. Fixed the imports as well
